+++ include/nstd/thread.h
@@ -23,11 +23,17 @@ typedef struct {
     NSTDAnyMut handle;
 } NSTDThreadHandle;
 
+/// Represents an optional value of type `NSTDThreadHandle`.
+NSTDOptional(NSTDThreadHandle) NSTDOptionalThreadHandle;
+
 /// A thread's unique identifier.
 typedef struct {
     /// The thread ID.
     NSTDAnyMut id;
 } NSTDThreadID;
+
+/// Represents an optional value of type `NSTDThreadID`.
+NSTDOptional(NSTDThreadID) NSTDOptionalThreadID;
 
 /// Describes the creation of a new thread.
 ///
@@ -84,8 +90,9 @@ NSTDAPI NSTDOptionalThread nstd_thread_spawn(
 ///
 /// # Returns
 ///
-/// `NSTDThreadHandle handle` - A handle to the current thread.
-NSTDAPI NSTDThreadHandle nstd_thread_current(void);
+/// `NSTDOptionalThreadHandle handle` - A handle to the current thread on success, or an
+/// uninitialized "none" variant on error.
+NSTDAPI NSTDOptionalThreadHandle nstd_thread_current(void);
 
 /// Retrieves a raw handle to a thread.
 ///
@@ -95,8 +102,9 @@ NSTDAPI NSTDThreadHandle nstd_thread_current(void);
 ///
 /// # Returns
 ///
-/// `NSTDThreadHandle handle` - A raw handle to the thread.
-NSTDAPI NSTDThreadHandle nstd_thread_handle(const NSTDThread *thread);
+/// `NSTDOptionalThreadHandle handle` - A raw handle to the thread on success, or an uninitialized
+/// "none" variant on error.
+NSTDAPI NSTDOptionalThreadHandle nstd_thread_handle(const NSTDThread *thread);
 
 /// Checks if a thread has finished running.
 ///
@@ -151,8 +159,9 @@ NSTDAPI NSTDOptionalStr nstd_thread_name(const NSTDThreadHandle *handle);
 ///
 /// # Returns
 ///
-/// `NSTDThreadID id` - The thread's unique ID.
-NSTDAPI NSTDThreadID nstd_thread_id(const NSTDThreadHandle *handle);
+/// `NSTDOptionalThreadID id` - The thread's unique ID on success, or an uninitialized "none"
+/// variant on error.
+NSTDAPI NSTDOptionalThreadID nstd_thread_id(const NSTDThreadHandle *handle);
 
 /// Frees an instance of `NSTDThreadHandle`.
 ///