+++ src/io/stderr.rs
@@ -1,6 +1,10 @@
 //! A handle to the standard error stream.
 use crate::{
-    core::slice::NSTDSlice,
+    alloc::CBox,
+    core::{
+        optional::{gen_optional, NSTDOptional},
+        slice::NSTDSlice,
+    },
     io::{NSTDIOError, NSTDIOResult},
 };
 use nstdapi::nstdapi;
@@ -12,19 +16,22 @@ use std::os::unix::io::AsRawFd;
 #[nstdapi]
 pub struct NSTDStderr {
     /// Rust's [Stderr].
-    err: Box<Stderr>,
+    err: CBox<Stderr>,
 }
+gen_optional!(NSTDOptionalStderr, NSTDStderr);
 
 /// Constructs a new handle to the standard error stream.
 ///
 /// # Returns
 ///
-/// `NSTDStderr handle` - A handle to the standard error stream.
+/// `NSTDOptionalStderr handle` - A handle to the standard error stream, or an uninitialized "none"
+/// variant on error.
 #[inline]
 #[nstdapi]
-pub fn nstd_io_stderr() -> NSTDStderr {
-    NSTDStderr {
-        err: Box::new(std::io::stderr()),
+pub fn nstd_io_stderr() -> NSTDOptionalStderr {
+    match CBox::new(std::io::stderr()) {
+        Some(err) => NSTDOptional::Some(NSTDStderr { err }),
+        _ => NSTDOptional::None,
     }
 }
 
@@ -99,7 +106,7 @@ pub unsafe fn nstd_io_stderr_write_all(handle: &mut NSTDStderr, bytes: &NSTDSlic
 #[inline]
 #[nstdapi]
 pub fn nstd_io_stderr_flush(handle: &mut NSTDStderr) -> NSTDIOError {
-    crate::io::stdio::flush(&mut handle.err)
+    crate::io::stdio::flush(&mut *handle.err)
 }
 
 /// Frees an instance of `NSTDStderr`.
@@ -116,19 +123,22 @@ pub fn nstd_io_stderr_free(handle: NSTDStderr) {}
 #[nstdapi]
 pub struct NSTDStderrLock {
     /// Rust's [StderrLock].
-    err: Box<StderrLock<'static>>,
+    err: CBox<StderrLock<'static>>,
 }
+gen_optional!(NSTDOptionalStderrLock, NSTDStderrLock);
 
 /// Constructs a new locked handle to the standard error stream.
 ///
 /// # Returns
 ///
-/// `NSTDStderrLock handle` - A locked handle to the standard error stream.
+/// `NSTDOptionalStderrLock handle` - A locked handle to the standard error stream on success, or
+/// an uninitialized "none" variant on error.
 #[inline]
 #[nstdapi]
-pub fn nstd_io_stderr_lock() -> NSTDStderrLock {
-    NSTDStderrLock {
-        err: Box::new(std::io::stderr().lock()),
+pub fn nstd_io_stderr_lock() -> NSTDOptionalStderrLock {
+    match CBox::new(std::io::stderr().lock()) {
+        Some(err) => NSTDOptional::Some(NSTDStderrLock { err }),
+        _ => NSTDOptional::None,
     }
 }
 
@@ -209,7 +219,7 @@ pub unsafe fn nstd_io_stderr_lock_write_all(
 #[inline]
 #[nstdapi]
 pub fn nstd_io_stderr_lock_flush(handle: &mut NSTDStderrLock) -> NSTDIOError {
-    crate::io::stdio::flush(&mut handle.err)
+    crate::io::stdio::flush(&mut *handle.err)
 }
 
 /// Frees and unlocks an instance of `NSTDStderrLock`.